Show different commands to moderators and regular group members. Minor cleanup.
