+++ skills/links.js
@@ -1,4 +1,5 @@
 /*********************************************************************************
+
 Links to projects should be posted to the #projects channel, not #general.
 
 *********************************************************************************/
